AI_Interviewer/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Werkzeug==2.3.6
python-dotenv==1.0.1
boto3==1.34.31
botocore==1.34.31
requests==2.31.0
mysql-connector-python==8.3.0

python-dateutil==2.8.2
PyJWT==2.8.0
openai==1.58.1
httpx==0.27.2
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
tenacity==8.2.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2